stock_tracker/requirements.txt
Flask
yfinance
plotly
textblob
GoogleNews
python-dotenv
